Add resizable vertical layout
Horizontal tiled layout is useless in tall screen on X200T. This is not
a perfect solution since xmonad does not provide a vertical layout in
xmonad-contrib. It is implemented by using ResizableTall layout with three
master windows, and adjust heights by MirrorShrink.

Note that, Reflect layout only reflect a layout horizontally or
vertically, not "flip" the layout 90 degree.